requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
requests
python-dotenv
groq
langchain-text-splitters